Fixed String state parsing
